+++ src/theia/sfm/verify_two_view_matches.h
@@ -61,7 +61,7 @@ bool VerifyTwoViewMatches(
     const CameraIntrinsicsPrior& intrinsics2,
     const std::vector<FeatureCorrespondence>& correspondences,
     TwoViewInfo* twoview_info,
-    std::vector<int>* inlier_indices);
+    std::vector<size_t>* inlier_indices);
 
 }  // namespace theia
